Add files via upload
This is the second capstone I worked on during my time at Tech Elevator and it was done with my classmate Yujoong Kim. The goal was to demonstrate our knowledge by writing a RESTful API server and command-line application, which interacts with a designated database. Tech Elevator provided some starting code, including the login/register user functionality. 

The capstone met the requirements of the project, but still needs some work to be done. I plan on adding updates to it and will add them here at a later date.